* aggiornato configure.ac in modo da supportare alcuni argomenti di ingresso:   --enable-profile : abilita il profiling   --enable-debug   : abilita il debug con o0 Wall Werror pedantic   --enable-release : abilita  O3 architettura nativa pipe * aggiornato build.sh per supportare le nuove opzioni * rimosse virgole inutili da config.h ( la modalit? pedantic e werror le segnava   come errore   (attualmente ci sono tanti altri errori e non compila in --enable-debug)

+++ src/simulation/general_tests.cpp
@@ -101,7 +101,7 @@ bool general_tests::LoadFile(individual_map* pool)
         if (p_buffer != "in") //!!!!
         {
             string block = "";
-            for (int i = 0; i < p_buffer.size(); i++)
+            for (int i = 0; i < (int) p_buffer.size(); i++)
                 block += "!";
 
             if (p_buffer == block)
